move TransactionLogMessage to hapi-fhir-rest-server bundle (as per feedback)
